requirements.txt
docling-core
gradio
mlx-vlm
pillow
requests
beautifulsoup4